Button for live logging + cleanup
