+++ src/can/mod.rs
@@ -2,5 +2,4 @@ pub mod bus_load;
 pub mod driver;
 pub mod state;
 pub mod thread;
-#[cfg(test)]
-mod thread_tests;
+pub mod daq_logger;
