+++ src/main.rs
@@ -36,7 +36,7 @@ fn main() -> eframe::Result<()> {
     }
 
     let _can_thread =
-        can::thread::start_can_thread(can_to_ui_tx, ui_to_can_rx, settings.selected_source.clone());
+        can::thread::start_can_thread(can_to_ui_tx, ui_to_can_rx, settings.selected_source.clone(), settings.log_folder.clone());
 
     let per_img = eframe::icon_data::from_png_bytes(assets::PER_LOGO_BYTES)
         .expect("Failed to load logo image");
